refactor(main): group skills and plugins modules
