+++ src/main/skills/skillManager.ts
@@ -6,16 +6,16 @@ import fs from 'fs';
 import yaml from 'js-yaml';
 import path from 'path';
 
-import { ComputerUseSkillId } from '../shared/computerUse/constants';
-import { isComputerUseKitInstalled } from './computerUse/computerUseKit';
-import { cpRecursiveSync } from './fsCompat';
-import { t } from './i18n';
-import { getElectronNodeRuntimePath } from './libs/coworkUtil';
-import { resolveNodeRuntimeForSpawn } from './libs/nodeRuntime';
-import { appendPythonRuntimeToEnv } from './libs/pythonRuntime';
-import { mergeReports,scanMultipleSkillDirs } from './libs/skillSecurity/skillSecurityScanner';
-import type { SecurityReportAction,SkillSecurityReport } from './libs/skillSecurity/skillSecurityTypes';
-import { SqliteStore } from './sqliteStore';
+import { ComputerUseSkillId } from '../../shared/computerUse/constants';
+import { isComputerUseKitInstalled } from '../computerUse/computerUseKit';
+import { cpRecursiveSync } from '../fsCompat';
+import { t } from '../i18n';
+import { getElectronNodeRuntimePath } from '../libs/coworkUtil';
+import { resolveNodeRuntimeForSpawn } from '../libs/nodeRuntime';
+import { appendPythonRuntimeToEnv } from '../libs/pythonRuntime';
+import { mergeReports,scanMultipleSkillDirs } from '../libs/skillSecurity/skillSecurityScanner';
+import type { SecurityReportAction,SkillSecurityReport } from '../libs/skillSecurity/skillSecurityTypes';
+import { SqliteStore } from '../sqliteStore';
 
 /**
  * Resolve the user's login shell PATH on macOS/Linux.
